// src/guide/types/subtitle.ts
export enum TSubtitle_ru {
    CTE = 'Строительно-техническая экспертиза',
    engineeringCommunication = 'Инженерные коммуникации',
    safety = 'Безопасность',
    territory = 'Объекты на территории',
    parking = 'Парковка',
    window = 'Вид из окон',
    furniture = 'Мебель',
    class = 'Класс жилья',
    decorating = 'Отделка',
    isNull = 'null'
}

export enum TSubtitle_en {
    CTE = 'Construction and technical expertise',
    engineeringCommunication = 'Engineering Communication',
    safety = 'Safety',
    territory = 'Objects on the territory',
    parking = 'Parking',
    window = 'Window',
    furniture = 'Furniture',
    class = 'Class',
    decorating = 'Decorating',
    isNull = 'null'
}